Share fetchAddonsData and fetchTagsData via @repo/utils
Three near-identical implementations existed: addon-catalog had
non-paginated locals for both, while frontpage's sitemap inlined a
paginated copy of each. Consolidate on the paginated form in
@repo/utils/fetch-addons-query and update all call sites. addon-catalog
now paginates tag fetching too, which is more correct (server defaults
were truncating long lists).

Resolves the two "Move this to a shared location, @repo/utils?" TODOs
in apps/frontpage/app/addons/sitemap.ts.
