+++ apps/addon-catalog/app/(home)/tag/[...name]/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from 'next/navigation';
 // eslint-disable-next-line -- the variable is camelCase
 import { unstable_cache } from 'next/cache';
 import { Preview } from '../../../../components/preview';
-import { fetchTagsData } from '../../../../lib/fetch-tags-data';
+import { fetchTagsData } from '@repo/utils/fetch-addons-query';
 import { fetchTagDetailsData } from '../../../../lib/fetch-tag-details-data';
 import type { Tag } from '../../../../types';
